refactor: split oversized modules (security, hierarchical_memory)
Split security.rs (949 LOC) into security/ directory:
- mod.rs: SecureMemoryBackend struct
- query_validator.rs: SQL validation
- credential_scrubber.rs: credential redaction
- scope_enforcer.rs: access control
- tests.rs: all security tests

Further split hierarchical_memory/mod.rs (2103 LOC):
- experience_ops.rs: add/query experience methods
- pattern_ops.rs: pattern evolution
- consolidation.rs: memory consolidation
- tests.rs: all hierarchical memory tests

All source modules now under 500 LOC philosophy limit.

Closes #30
